Further fixes, the creature conditions now save & load. But combined conditions and injuries don't save/load correctly.
